Remove unused context and add get_guild_settings method
